Use underscore prefix for protected members
The library is written in Typescript where we don't need to follow
the convention for private members to use an underscore prefix. But it
is transpiled to JS and used from other JS code in editors that might
not look at type definitions. To them, these methods look like public
and I've seen people starting to use them.