Try next available list command after any FTP error (#117)
If a list command is not supported we want to try the next available
one. An FTP server should report unsupported commands with an error
code 5xx but it's possible that it uses 4xx.

So far the library wanted to pass along errors that are not related
to an unsupported command. But this can't be done reliably.